+++ EmpiriaBMS.Front/ViewModel/Components/OtherVM.cs
@@ -31,8 +31,8 @@ public class OtherVM : BaseVM
         }
     }
 
-    private int _completionEstimation;
-    public int CompletionEstimation
+    private float _completionEstimation;
+    public float CompletionEstimation
     {
         get => _completionEstimation;
         set
