+++ EmpiriaBMS.Models/Models/Draw.cs
@@ -17,7 +17,7 @@ public class Draw : Entity
 
     public double MenHours { get; set; }
 
-    public int CompletionEstimation { get; set; }
+    public float CompletionEstimation { get; set; }
 
     [DataType(DataType.DateTime)]
     [Column(TypeName = "datetime2")]
